requirements.txt
django==4.1.2
django-phonenumber-field[phonenumbers]
fontawesomefree==6
